Add helper class for CopyToTexture test suites
