+++ src/webgpu/util/copyToTexture.ts
@@ -1,0 +1,126 @@
+export const description = `
+copyToTexture with HTMLCanvasElement and OffscreenCanvas sources.
+
+- x= {HTMLCanvasElement, OffscreenCanvas}
+- x= {2d, webgl, webgl2, gpupresent, bitmaprenderer} context, {various context creation attributes}
+
+TODO: consider whether external_texture and copyToTexture video tests should be in the same file
+TODO: plan
+`;
+
+import { GPUTest } from '../gpu_test.js';
+
+function calculateRowPitch(width: number, bytesPerPixel: number): number {
+  const bytesPerRow = width * bytesPerPixel;
+  // Rounds up to a multiple of 256 according to WebGPU requirements.
+  return (((bytesPerRow - 1) >> 8) + 1) << 8;
+}
+
+export class CopyToTextureUtils extends GPUTest {
+  checkCopyExternalImageResult(
+    src: GPUBuffer,
+    expected: ArrayBufferView,
+    width: number,
+    height: number,
+    bytesPerPixel: number
+  ): void {
+    const exp = new Uint8Array(expected.buffer, expected.byteOffset, expected.byteLength);
+    const rowPitch = calculateRowPitch(width, bytesPerPixel);
+    const dst = this.createCopyForMapRead(src, 0, rowPitch * height);
+
+    this.eventualAsyncExpectation(async niceStack => {
+      await dst.mapAsync(GPUMapMode.READ);
+      const actual = new Uint8Array(dst.getMappedRange());
+      const check = this.checkBufferWithRowPitch(
+        actual,
+        exp,
+        width,
+        height,
+        rowPitch,
+        bytesPerPixel
+      );
+      if (check !== undefined) {
+        niceStack.message = check;
+        this.rec.expectationFailed(niceStack);
+      }
+      dst.destroy();
+    });
+  }
+
+  checkBufferWithRowPitch(
+    actual: Uint8Array,
+    exp: Uint8Array,
+    width: number,
+    height: number,
+    rowPitch: number,
+    bytesPerPixel: number
+  ): string | undefined {
+    const failedByteIndices: string[] = [];
+    const failedByteExpectedValues: string[] = [];
+    const failedByteActualValues: string[] = [];
+    iLoop: for (let i = 0; i < height; ++i) {
+      const bytesPerRow = width * bytesPerPixel;
+      for (let j = 0; j < bytesPerRow; ++j) {
+        const indexExp = j + i * bytesPerRow;
+        const indexActual = j + rowPitch * i;
+        if (actual[indexActual] !== exp[indexExp]) {
+          if (failedByteIndices.length >= 4) {
+            failedByteIndices.push('...');
+            failedByteExpectedValues.push('...');
+            failedByteActualValues.push('...');
+            break iLoop;
+          }
+          failedByteIndices.push(`(${i},${j})`);
+          failedByteExpectedValues.push(exp[indexExp].toString());
+          failedByteActualValues.push(actual[indexActual].toString());
+        }
+      }
+    }
+    if (failedByteIndices.length > 0) {
+      return `at [${failedByteIndices.join(', ')}], \
+expected [${failedByteExpectedValues.join(', ')}], \
+got [${failedByteActualValues.join(', ')}]`;
+    }
+    return undefined;
+  }
+
+  doTestAndCheckResult(
+    imageCopyExternalImage: GPUImageCopyExternalImage,
+    dstTextureCopyView: GPUImageCopyTexture,
+    copySize: GPUExtent3DDict,
+    bytesPerPixel: number,
+    expectedData: Uint8ClampedArray
+  ): void {
+    this.device.queue.copyExternalImageToTexture(
+      imageCopyExternalImage,
+      dstTextureCopyView,
+      copySize
+    );
+
+    const externalImage = imageCopyExternalImage.source;
+    const dstTexture = dstTextureCopyView.texture;
+
+    const bytesPerRow = calculateRowPitch(externalImage.width, bytesPerPixel);
+    const testBuffer = this.device.createBuffer({
+      size: bytesPerRow * externalImage.height,
+      usage: GPUBufferUsage.COPY_SRC | GPUBufferUsage.COPY_DST,
+    });
+
+    const encoder = this.device.createCommandEncoder();
+
+    encoder.copyTextureToBuffer(
+      { texture: dstTexture, mipLevel: 0, origin: { x: 0, y: 0, z: 0 } },
+      { buffer: testBuffer, bytesPerRow },
+      { width: externalImage.width, height: externalImage.height, depthOrArrayLayers: 1 }
+    );
+    this.device.queue.submit([encoder.finish()]);
+
+    this.checkCopyExternalImageResult(
+      testBuffer,
+      expectedData,
+      externalImage.width,
+      externalImage.height,
+      bytesPerPixel
+    );
+  }
+}
